requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
dj-database-url==2.0.0
Django==4.2.1
django-bootstrap5==23.1
django-heroku==0.3.1
Faker==18.7.0
gunicorn==20.1.0
idna==3.4
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
whitenoise==6.4.0
